third_party/blink/web_tests/http/tests/security/subresourceIntegrity/subresource-integrity-script-cors.html

<!DOCTYPE html>
<head>
    <title>SRI on cross origin CORS enabled script</title>
</head>
<body>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        setup({single_test: true});

        var result = '';
        var scriptLoaded = function() {
            assert_equals(result, 'ran', 'Script ran.');
            done();
        };
        var scriptError = function() {
            assert_unreached('Script error.');
            done();
        };
    </script>
    <script crossorigin integrity="sha256-mk_DwEwGGoyrh4sgi9t0nYTO1Lrft3w6vzaNuMaZrtQ=" src="http://localhost:8000/security/resources/cors-script.php?value=ran" onload="scriptLoaded();" onerror="scriptError();"></script>
</body>
</html>
